Separate rallyme script functions from library code
