+++ scripts/rallyme.php
@@ -1,9 +1,7 @@
 <?
-require('include/slack.php');
-require('include/curl.php');
-require('include/rally.php');
 require('include/rallyme.config.php');
 require('include/slack.config.php');
+require('include/rallyme.inc.php');
 
 $slackCommand = BuildSlashCommand($_REQUEST);
